fix: remove credential page, fix get by id
