+++ oid4vci/demo/frontend/src/App.js
@@ -1,7 +1,6 @@
 import React from 'react';
 import './App.css';
 import { BrowserRouter as Router, Route, Routes } from 'react-router-dom';
-import FormPage from './FormPage';
 import InputForm from './InputForm';
 import RegistrationPage from './AdminPage'
 import QRCodePage from './QRCodePage';
@@ -13,7 +12,6 @@ function App() {
         <Routes>
           <Route path="/" exact element={<RegistrationPage/>} />
           <Route path="/input" exact element={<InputForm/>} />
-          <Route path="/credentials" element={<FormPage/>} />
           <Route path="/qr-code" element={<QRCodePage/>} />
         </Routes>
       </div>
