+++ oid4vci/demo/frontend/src/InputForm.js
@@ -1,4 +1,5 @@
 import React, { useState } from 'react';
+import axios from "axios";
 import { useNavigate, useLocation } from "react-router-dom";
 
 
@@ -24,9 +25,53 @@ const InputForm = () => {
     setEmail(e.target.value);
   };
 
-  const handleShareClick = () => {
-    navigate(`/credentials`,{ state: {firstName:firstName, lastName:lastName, email:email, did:did, credential:supportedCredId}});
-    
+  const getOffer = () => {
+    // Set the Axios configuration for CORS and credentials
+    axios.defaults.withCredentials = true; // Enable credentials (cookies, etc.)
+    axios.defaults.headers.common["Access-Control-Allow-Origin"] =
+      "http://localhost:3001"; // Adjust the origin as needed
+    console.log(firstName, lastName, email, did);
+
+    // api call to controller, `POST /exchange/submit`
+    axios
+      .post("http://localhost:3001/oid4vci/exchange/create", {
+        credential_subject: {
+          name: firstName,
+          lastname: lastName,
+          email,
+        },
+        did: did,
+        supported_cred_id: supportedCredId,
+      })
+      .then((response) => {
+        console.log(response.data);
+        const { exchange_id } = response.data;
+        // TODO: call offer endpoint
+
+        const queryParams = {
+          user_pin_required: false,
+          exchange_id: exchange_id,
+        };
+        console.log("get offer params:");
+        console.log(queryParams);
+        axios
+          .get("http://localhost:3001/oid4vci/credential-offer", {
+            params: queryParams,
+            headers: {
+              accept: "application/json",
+            },
+          })
+          .then((response) => {
+            console.log(response.data);
+            const credentialOffer = response.data;
+            navigate(`/qr-code`, {
+              state: { credentialOffer, exchange_id: exchange_id },
+            });
+          });
+      })
+      .catch((error) => {
+        console.error(error);
+      });
   };
 
   return (
@@ -61,7 +106,7 @@ const InputForm = () => {
               onChange={handleEmailChange}
             />
           </div>
-        <button type="button" onClick={handleShareClick}>
+        <button type="button" onClick={getOffer}>
           Share
         </button>
       </form>
